refactor(public): move public to root directory
move public to root directory, because vite looking for public in the root directory as default

remove static directory because vite copy files to build directory from public directory when building project
