+++ src/components/Avatar/Avatar.scss
@@ -5,7 +5,7 @@
   margin: 0 auto;
   border-radius: 100%;
   background-color: $gray-20;
-  background-image: url('../../public/images/svg/empty.svg');
+  background-image: url('/images/svg/empty.svg');
   background-repeat: no-repeat;
   background-position: center center;
   margin-bottom: 90px;
