+++ src/components/SignUp/index.ts
@@ -1,3 +1,3 @@
-import '../../public/styles/loginPage.scss'
+import '../../styles/loginPage.scss'
 
 export { default as SignUp } from './SignUp'
